+++ CHANGELOG.md
@@ -3,8 +3,9 @@ Changelog
 
 v2.3.0
 ------
-* Fix potential race condition which can cause the ClassLoader to change.
+* Fix a potential race condition which can cause the ClassLoader to change.
 * Fix a bug which ignored clinit exception handlers for all but the first one.
+* Define a new method for creating explicitly named classes.
 * Add a convenience method which checks if an added method overrides an inherited one.
 
 v2.2.3 (2022-05-15)